Add discrete things

+++ src/levels/level.rs
@@ -62,10 +62,10 @@ pub struct EscalatorConfig {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ThingConfig {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i32,
+    pub y: i32,
+    pub width: i32,
+    pub height: i32,
     pub color_flag: ColorFlag,
 }
